# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for simpo pharmacy management system supporting POS, inventory, reporting, and multi-branch operations.",
        "title": "simpo Pharmacy Management System API",
        "termsOfService": "https://simpo.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@simpo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/admin/users/{user_id}/permissions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Grants a specific permission to a user with audit logging (Story 6.4, AC1, AC2)\nNote: Current implementation uses RBAC with pre-defined roles. This endpoint\nis provided for future PBAC (Permission-Based Access Control) enhancement.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Management"
                ],
                "summary": "Grant permission to user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Permission grant request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.GrantPermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission granted successfully",
                        "schema": {
                            "$ref": "#/definitions/user.GrantPermissionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - SYSTEM_ADMIN only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "501": {
                        "description": "Not Implemented - RBAC system in use",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revokes a specific permission from a user with audit logging (Story 6.4, AC1, AC2)\nNote: Current implementation uses RBAC with pre-defined roles. This endpoint\nis provided for future PBAC (Permission-Based Access Control) enhancement.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Management"
                ],
                "summary": "Revoke permission from user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Permission revoke request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.RevokePermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission revoked successfully",
                        "schema": {
                            "$ref": "#/definitions/user.RevokePermissionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - SYSTEM_ADMIN only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "501": {
                        "description": "Not Implemented - RBAC system in use",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/users/{user_id}/role": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a user's role with audit logging (Story 6.4, AC1, AC2)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Management"
                ],
                "summary": "Update user role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role updated successfully",
                        "schema": {
                            "$ref": "#/definitions/user.UpdateRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - SYSTEM_ADMIN only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "user.GrantPermissionRequest": {
            "type": "object",
            "required": [
                "permission",
                "reason"
            ],
            "properties": {
                "permission": {
                    "description": "Permission is the permission to grant\nExamples: \"MANAGE_USERS\", \"MANAGE_INVENTORY\", \"VIEW_REPORTS\"",
                    "type": "string",
                    "minLength": 3,
                    "example": "MANAGE_USERS"
                },
                "reason": {
                    "description": "Reason is the reason for granting permission (minimum 5 characters)\nExample: \"Additional responsibilities for new project\"",
                    "type": "string",
                    "minLength": 5,
                    "example": "Additional responsibilities for new project"
                }
            }
        },
        "user.GrantPermissionResponse": {
            "type": "object",
            "properties": {
                "granted_at": {
                    "description": "GrantedAt is the timestamp when permission was granted\nExample: \"2026-05-27T12:00:00Z\"",
                    "type": "string",
                    "example": "2026-05-27T12:00:00Z"
                },
                "id": {
                    "description": "ID is the unique identifier for the user\nExample: 10",
                    "type": "integer",
                    "example": 10
                },
                "permission": {
                    "description": "Permission is the granted permission\nExample: \"MANAGE_USERS\"",
                    "type": "string",
                    "example": "MANAGE_USERS"
                },
                "username": {
                    "description": "Username is the login username\nExample: \"staffuser\"",
                    "type": "string",
                    "example": "staffuser"
                }
            }
        },
...
        "user.RevokePermissionRequest": {
            "type": "object",
            "required": [
                "permission",
                "reason"
            ],
            "properties": {
                "permission": {
                    "description": "Permission is the permission to revoke\nExamples: \"MANAGE_USERS\", \"MANAGE_INVENTORY\", \"VIEW_REPORTS\"",
                    "type": "string",
                    "minLength": 3,
                    "example": "MANAGE_USERS"
                },
                "reason": {
                    "description": "Reason is the reason for revoking permission (minimum 5 characters)\nExample: \"Project completed, no longer needs access\"",
                    "type": "string",
                    "minLength": 5,
                    "example": "Project completed, no longer needs access"
                }
            }
        },
        "user.RevokePermissionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID is the unique identifier for the user\nExample: 10",
                    "type": "integer",
                    "example": 10
                },
                "permission": {
                    "description": "Permission is the revoked permission\nExample: \"MANAGE_USERS\"",
                    "type": "string",
                    "example": "MANAGE_USERS"
                },
                "revoked_at": {
                    "description": "RevokedAt is the timestamp when permission was revoked\nExample: \"2026-05-27T12:00:00Z\"",
                    "type": "string",
                    "example": "2026-05-27T12:00:00Z"
                },
                "username": {
                    "description": "Username is the login username\nExample: \"staffuser\"",
                    "type": "string",
                    "example": "staffuser"
                }
            }
        },
...
        "user.UpdateRoleRequest": {
            "type": "object",
            "required": [
                "reason",
                "role"
            ],
            "properties": {
                "reason": {
                    "description": "Reason is the reason for role change (minimum 5 characters)\nExamples: \"Promoted to owner\", \"Role change due to reorganization\", \"Demoted for policy violation\"",
                    "type": "string",
                    "minLength": 5,
                    "example": "Promoted to owner position"
                },
                "role": {
                    "description": "Role is the new role to assign (must be one of: SYSTEM_ADMIN, OWNER, CASHIER)\nExample: \"OWNER\"\nEnum: SYSTEM_ADMIN, OWNER, CASHIER",
                    "type": "string",
                    "example": "OWNER"
                }
            }
        },
        "user.UpdateRoleResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID is the unique identifier for the user\nExample: 10",
                    "type": "integer",
                    "example": 10
                },
                "new_role": {
                    "description": "NewRole is the new role after change\nExample: \"OWNER\"",
                    "type": "string",
                    "example": "OWNER"
                },
                "old_role": {
                    "description": "OldRole is the previous role before change\nExample: \"CASHIER\"",
                    "type": "string",
                    "example": "CASHIER"
                },
                "updated_at": {
                    "description": "UpdatedAt is the timestamp when role was changed\nExample: \"2026-05-27T12:00:00Z\"",
                    "type": "string",
                    "example": "2026-05-27T12:00:00Z"
                },
                "username": {
                    "description": "Username is the login username\nExample: \"staffuser\"",
                    "type": "string",
                    "example": "staffuser"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}